feat(hero): Add hero section
Add html structure for hero section. Comment out hero img while working on mobile view as it only appears at min width of 768px. Also added .social-links to .navbar__social-links and social-link to navbar__social-link. Social links appear on both the menu and the hero section, it will be better to make it as a global class.